Fix clashing namespace.
