+++ osu.Game.Modes.Taiko/Objects/Drawables/DrawableCentreHit.cs
@@ -1,18 +1,18 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using OpenTK.Input;
-using osu.Game.Modes.Taiko.Objects.Drawable.Pieces;
 using osu.Framework.Allocation;
 using osu.Game.Graphics;
+using osu.Game.Modes.Taiko.Objects.Drawables.Pieces;
+using OpenTK.Input;
 
-namespace osu.Game.Modes.Taiko.Objects.Drawable
+namespace osu.Game.Modes.Taiko.Objects.Drawables
 {
-    public class DrawableStrongCentreHit : DrawableStrongHit
+    public class DrawableCentreHit : DrawableHit
     {
         protected override Key[] HitKeys { get; } = { Key.F, Key.J };
 
-        public DrawableStrongCentreHit(Hit hit)
+        public DrawableCentreHit(Hit hit)
             : base(hit)
         {
         }
@@ -23,7 +23,7 @@ namespace osu.Game.Modes.Taiko.Objects.Drawable
             Circle.AccentColour = colours.PinkDarker;
         }
 
-        protected override CirclePiece CreateCirclePiece() => new StrongCirclePiece
+        protected override CirclePiece CreateCirclePiece() => new CirclePiece
         {
             Children = new[] { new CentreHitSymbolPiece() }
         };
