+++ osu.Game.Modes.Taiko/Objects/Drawables/DrawableHit.cs
@@ -1,16 +1,16 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using OpenTK.Input;
+using System;
+using System.Linq;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Modes.Objects.Drawables;
 using osu.Game.Modes.Taiko.Judgements;
-using osu.Game.Modes.Taiko.Objects.Drawable.Pieces;
-using System;
-using System.Linq;
+using osu.Game.Modes.Taiko.Objects.Drawables.Pieces;
+using OpenTK.Input;
 
-namespace osu.Game.Modes.Taiko.Objects.Drawable
+namespace osu.Game.Modes.Taiko.Objects.Drawables
 {
     public abstract class DrawableHit : DrawableTaikoHitObject
     {
@@ -19,7 +19,7 @@ namespace osu.Game.Modes.Taiko.Objects.Drawable
         /// </summary>
         protected abstract Key[] HitKeys { get; }
 
-        protected override Container<Framework.Graphics.Drawable> Content => bodyContainer;
+        protected override Container<Drawable> Content => bodyContainer;
 
         protected readonly CirclePiece Circle;
