+++ osu.Game.Modes.Taiko/Objects/Drawables/DrawableMajorBarLine.cs
@@ -6,7 +6,7 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
 using OpenTK;
 
-namespace osu.Game.Modes.Taiko.Objects.Drawable
+namespace osu.Game.Modes.Taiko.Objects.Drawables
 {
     public class DrawableMajorBarLine : DrawableBarLine
     {
