+++ osu.Game.Modes.Taiko/Objects/Drawables/DrawableRimHit.cs
@@ -1,18 +1,18 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using OpenTK.Input;
-using osu.Game.Modes.Taiko.Objects.Drawable.Pieces;
-using osu.Game.Graphics;
 using osu.Framework.Allocation;
+using osu.Game.Graphics;
+using osu.Game.Modes.Taiko.Objects.Drawables.Pieces;
+using OpenTK.Input;
 
-namespace osu.Game.Modes.Taiko.Objects.Drawable
+namespace osu.Game.Modes.Taiko.Objects.Drawables
 {
-    public class DrawableCentreHit : DrawableHit
+    public class DrawableRimHit : DrawableHit
     {
-        protected override Key[] HitKeys { get; } = { Key.F, Key.J };
+        protected override Key[] HitKeys { get; } = { Key.D, Key.K };
 
-        public DrawableCentreHit(Hit hit)
+        public DrawableRimHit(Hit hit)
             : base(hit)
         {
         }
@@ -20,12 +20,12 @@ namespace osu.Game.Modes.Taiko.Objects.Drawable
         [BackgroundDependencyLoader]
         private void load(OsuColour colours)
         {
-            Circle.AccentColour = colours.PinkDarker;
+            Circle.AccentColour = colours.BlueDarker;
         }
 
         protected override CirclePiece CreateCirclePiece() => new CirclePiece
         {
-            Children = new[] { new CentreHitSymbolPiece() }
+            Children = new[] { new RimHitSymbolPiece() }
         };
     }
 }
