+++ osu.Game.Modes.Taiko/osu.Game.Modes.Taiko.csproj
@@ -54,25 +54,25 @@
     <Compile Include="Judgements\TaikoJudgement.cs" />
     <Compile Include="Judgements\TaikoHitResult.cs" />
     <Compile Include="Objects\BarLine.cs" />
-    <Compile Include="Objects\Drawable\DrawableBarLine.cs" />
-    <Compile Include="Objects\Drawable\DrawableMajorBarLine.cs" />
+    <Compile Include="Objects\Drawables\DrawableBarLine.cs" />
+    <Compile Include="Objects\Drawables\DrawableMajorBarLine.cs" />
     <Compile Include="Objects\CentreHit.cs" />
-    <Compile Include="Objects\Drawable\DrawableRimHit.cs" />
-    <Compile Include="Objects\Drawable\DrawableStrongRimHit.cs" />
-    <Compile Include="Objects\Drawable\DrawableCentreHit.cs" />
-    <Compile Include="Objects\Drawable\DrawableHit.cs" />
-    <Compile Include="Objects\Drawable\DrawableStrongDrumRoll.cs" />
-    <Compile Include="Objects\Drawable\DrawableStrongCentreHit.cs" />
-    <Compile Include="Objects\Drawable\DrawableStrongHit.cs" />
-    <Compile Include="Objects\Drawable\Pieces\CentreHitSymbolPiece.cs" />
-    <Compile Include="Objects\Drawable\DrawableDrumRoll.cs" />
-    <Compile Include="Objects\Drawable\DrawableDrumRollTick.cs" />
-    <Compile Include="Objects\Drawable\DrawableSwell.cs" />
-    <Compile Include="Objects\Drawable\DrawableTaikoHitObject.cs" />
-    <Compile Include="Objects\Drawable\Pieces\RimHitSymbolPiece.cs" />
-    <Compile Include="Objects\Drawable\Pieces\StrongCirclePiece.cs" />
-    <Compile Include="Objects\Drawable\Pieces\CirclePiece.cs" />
-    <Compile Include="Objects\Drawable\Pieces\SwellSymbolPiece.cs" />
+    <Compile Include="Objects\Drawables\DrawableRimHit.cs" />
+    <Compile Include="Objects\Drawables\DrawableStrongRimHit.cs" />
+    <Compile Include="Objects\Drawables\DrawableCentreHit.cs" />
+    <Compile Include="Objects\Drawables\DrawableHit.cs" />
+    <Compile Include="Objects\Drawables\DrawableStrongDrumRoll.cs" />
+    <Compile Include="Objects\Drawables\DrawableStrongCentreHit.cs" />
+    <Compile Include="Objects\Drawables\DrawableStrongHit.cs" />
+    <Compile Include="Objects\Drawables\Pieces\CentreHitSymbolPiece.cs" />
+    <Compile Include="Objects\Drawables\DrawableDrumRoll.cs" />
+    <Compile Include="Objects\Drawables\DrawableDrumRollTick.cs" />
+    <Compile Include="Objects\Drawables\DrawableSwell.cs" />
+    <Compile Include="Objects\Drawables\DrawableTaikoHitObject.cs" />
+    <Compile Include="Objects\Drawables\Pieces\RimHitSymbolPiece.cs" />
+    <Compile Include="Objects\Drawables\Pieces\StrongCirclePiece.cs" />
+    <Compile Include="Objects\Drawables\Pieces\CirclePiece.cs" />
+    <Compile Include="Objects\Drawables\Pieces\SwellSymbolPiece.cs" />
     <Compile Include="Objects\DrumRoll.cs" />
     <Compile Include="Objects\DrumRollTick.cs" />
     <Compile Include="Objects\Hit.cs" />
